requirements.txt
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1
requests==2.32.3
python-dateutil==2.9.0.post0
openai==1.51.2
